Parse the vertex buffers into Vertex arrays instead of keeping them as raw byte data
The offsets and lengths in the Mesh class are in units of Vertices, not bytes. Therefore, in order
to use the Mesh data, we need to parse the Vertex buffers into Vertex arrays.

First we classify each vertex buffer as holding either SIMPLE vertices or COMPLEX vertices. We
assume that the whole buffer uses the same type.

Then we read through each buffer using the classification to "cast" the byte array to a
Vertex array.